Get all incidents
